requirements.txt
pandas>=2.0.0
scikit-learn>=1.0.0
numpy>=1.20.0
google-generativeai>=0.3.0
matplotlib>=3.4.0